requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
pytz==2018.7
